requirements.txt
python-telegram-bot==20.7
aiohttp==3.8.6
psycopg2-binary==2.9.9
